[REFACTORING]: moved title and text components to the specific folder
