+++ v6y-apps/front-bo/src/commons/components/VitalityEmptyView.tsx
@@ -1,4 +1,4 @@
-import { VitalityText } from '@v6y/shared-ui/src/components/VitalityText';
+import { VitalityText } from '@v6y/shared-ui/src/components/VitalityText/VitalityText';
 import { Empty } from 'antd';
 import * as React from 'react';
