+++ v6y-apps/front-bo/src/features/v6y-accounts/components/VitalityAccountCreateView.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import { ApplicationType } from '@v6y/core-logic/src/types/ApplicationType';
-import { VitalityTitle } from '@v6y/shared-ui/src/components/VitalityTitle';
+import { VitalityTitle } from '@v6y/shared-ui/src/components/VitalityTitle/VitalityTitle';
 import * as React from 'react';
 import { useEffect, useState } from 'react';
