+++ v6y-apps/front-bo/src/features/v6y-auth/VitalityAuthForgotPasswordView.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import { AuthPage as AuthPageBase } from '@refinedev/antd';
-import { VitalityTitle } from '@v6y/shared-ui/src/components/VitalityTitle';
+import { VitalityTitle } from '@v6y/shared-ui/src/components/VitalityTitle/VitalityTitle';
 import React from 'react';
 
 export const VitalityAuthForgotPasswordView = () => {
