+++ v6y-apps/front-bo/src/features/v6y-deprecated-dependencies/components/VitalityDeprecatedDependencyDetailsView.tsx
@@ -1,6 +1,6 @@
 import { useParsed } from '@refinedev/core';
 import { DeprecatedDependencyType } from '@v6y/core-logic/src';
-import { VitalityTitle } from '@v6y/shared-ui/src/components/VitalityTitle';
+import { VitalityTitle } from '@v6y/shared-ui/src/components/VitalityTitle/VitalityTitle';
 import * as React from 'react';
 
 import VitalityDetailsView from '../../../commons/components/VitalityDetailsView';
