+++ v6y-apps/front-bo/src/features/v6y-evolution-helps/components/VitalityEvolutionHelpEditView.tsx
@@ -1,6 +1,6 @@
 import { useParsed } from '@refinedev/core';
 import { EvolutionHelpType } from '@v6y/core-logic';
-import { VitalityTitle } from '@v6y/shared-ui/src/components/VitalityTitle';
+import { VitalityTitle } from '@v6y/shared-ui/src/components/VitalityTitle/VitalityTitle';
 import * as React from 'react';
 
 import {
