+++ v6y-apps/front-bo/src/features/v6y-faqs/components/VitalityFaqCreateView.tsx
@@ -1,4 +1,4 @@
-import { VitalityTitle } from '@v6y/shared-ui/src/components/VitalityTitle';
+import { VitalityTitle } from '@v6y/shared-ui/src/components/VitalityTitle/VitalityTitle';
 import * as React from 'react';
 
 import {
