+++ v6y-apps/front-bo/src/features/v6y-faqs/components/VitalityFaqEditView.tsx
@@ -1,5 +1,5 @@
 import { useParsed } from '@refinedev/core';
-import { VitalityTitle } from '@v6y/shared-ui/src/components/VitalityTitle';
+import { VitalityTitle } from '@v6y/shared-ui/src/components/VitalityTitle/VitalityTitle';
 import * as React from 'react';
 
 import {
