+++ v6y-apps/front/src/commons/components/application-info/VitalityAppInfos.tsx
@@ -1,4 +1,4 @@
-import { VitalityText } from '@v6y/shared-ui/src/components/VitalityText';
+import { VitalityText } from '@v6y/shared-ui/src/components/VitalityText/VitalityText';
 import { Col, Divider, List, Row, Tag } from 'antd';
 import Link from 'next/link';
 import * as React from 'react';
