+++ v6y-apps/front/src/commons/components/layout/VitalityPageLayout.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { VitalityTitle } from '@v6y/shared-ui/src/components/VitalityTitle';
+import { VitalityTitle } from '@v6y/shared-ui/src/components/VitalityTitle/VitalityTitle';
 import { Col, FloatButton, Layout, Row } from 'antd';
 import * as React from 'react';
 import { ReactNode } from 'react';
