+++ v6y-apps/front/src/commons/components/modules/VitalityModulesView.tsx
@@ -1,6 +1,6 @@
 import { InfoCircleOutlined, PushpinOutlined } from '@ant-design/icons';
-import { VitalityText } from '@v6y/shared-ui/src/components/VitalityText';
-import { VitalityTitle } from '@v6y/shared-ui/src/components/VitalityTitle';
+import { VitalityText } from '@v6y/shared-ui/src/components/VitalityText/VitalityText';
+import { VitalityTitle } from '@v6y/shared-ui/src/components/VitalityTitle/VitalityTitle';
 import { Avatar, Button, Card, Divider, List, Space, Statistic } from 'antd';
 import * as React from 'react';
 import { useEffect, useState } from 'react';
