+++ v6y-apps/front/src/commons/hooks/useDataGrouper.tsx
@@ -1,4 +1,4 @@
-import { VitalityText } from '@v6y/shared-ui/src/components/VitalityText';
+import { VitalityText } from '@v6y/shared-ui/src/components/VitalityText/VitalityText';
 import * as React from 'react';
 import { useEffect, useState } from 'react';
