+++ v6y-apps/front/src/features/app-details/components/audit-reports/auditors/lighthouse/VitalityLighthouseReportsCategoryGrouper.tsx
@@ -1,7 +1,7 @@
 import { InfoCircleOutlined } from '@ant-design/icons';
 import { AuditType } from '@v6y/core-logic';
-import { VitalityText } from '@v6y/shared-ui/src/components/VitalityText';
-import { VitalityTitle } from '@v6y/shared-ui/src/components/VitalityTitle';
+import { VitalityText } from '@v6y/shared-ui/src/components/VitalityText/VitalityText';
+import { VitalityTitle } from '@v6y/shared-ui/src/components/VitalityTitle/VitalityTitle';
 import { Button, Card, Col, List, Row, Statistic } from 'antd';
 import Link from 'next/link';
 import * as React from 'react';
