+++ v6y-apps/front/src/features/app-list/components/VitalityAppListHeader.tsx
@@ -1,5 +1,5 @@
 import { ExportOutlined } from '@ant-design/icons';
-import { VitalityTitle } from '@v6y/shared-ui/src/components/VitalityTitle';
+import { VitalityTitle } from '@v6y/shared-ui/src/components/VitalityTitle/VitalityTitle';
 import { Button, Col, Row } from 'antd';
 import { useEffect, useState } from 'react';
 import * as React from 'react';
