+++ v6y-apps/front/src/features/dashboard/components/VitalityDashboardMenu.tsx
@@ -1,4 +1,4 @@
-import { VitalityTitle } from '@v6y/shared-ui/src/components/VitalityTitle';
+import { VitalityTitle } from '@v6y/shared-ui/src/components/VitalityTitle/VitalityTitle';
 import { Col, Row } from 'antd';
 import * as React from 'react';
