+++ v6y-libs/shared-ui/src/components/VitalityText/VitalityText.tsx
