+++ v6y-libs/shared-ui/src/index.ts
@@ -1,4 +1,4 @@
-import * as VitalityText from './components/VitalityText';
-import * as VitalityTitle from './components/VitalityTitle';
+import * as VitalityText from './components/VitalityText/VitalityText';
+import * as VitalityTitle from './components/VitalityTitle/VitalityTitle';
 
 export { VitalityText, VitalityTitle };
